requirements.txt
langchain
python-dotenv
streamlit
#uncomment to use openai
#openai
qdrant-client
tiktoken


# uncomment to use huggingface llms
huggingface-hub==0.14.1

# uncomment to use instructor embeddings
InstructorEmbedding==1.0.1
sentence-transformers==2.2.2